move stub display into a class
