+++ spec/game_spec.rb
@@ -1,26 +1,7 @@
 require "spec_helper"
 require "game"
 require "human_double"
-
-class StubDisplay
-  def show_board(_board)
-  end
-
-  def announce_tie
-  end
-  
-  def announce_win(player)
-  end
-
-  def clear_screen
-  end
-  
-  def prompt_player(player)
-  end
-
-  def notify_invalid(selection)
-  end
-end
+require "stub_display"
 
 RSpec.describe Game do
  it "knows that the game can continue" do
