+++ spec/stub_display.rb
@@ -1,0 +1,21 @@
+class StubDisplay
+  def show_board(_board)
+  end
+
+  def announce_tie
+  end
+  
+  def announce_win(player)
+  end
+
+  def clear_screen
+  end
+  
+  def prompt_player(player)
+  end
+
+  def notify_invalid(selection)
+  end
+end
+
+
